renamed old dropdown item to search item
